ESQL: Rename evaluators for FIRST and LAST
I'm going to be reusing the evaluators from FIRST_OVER_TIME and
LAST_OVER_TIME for FIRST and LAST. This renames the evaluators
themselves to indicate that they are the actual evaluators for FIRST and
LAST.

Relates to #108385
