+++ x-pack/plugin/esql/compute/build.gradle
@@ -900,28 +900,28 @@ tasks.named('stringTemplates').configure {
   }
 
   // TODO: add {value}_over_time for other types: boolean, bytes_refs
-  File valueOverTimeAggregatorInputFile = file("src/main/java/org/elasticsearch/compute/aggregation/X-ValueOverTimeAggregator.java.st")
+  File valueByLongAggregatorInputFile = file("src/main/java/org/elasticsearch/compute/aggregation/X-ValueByLongAggregator.java.st")
   ["First", "Last"].forEach { Occurrence ->
     {
       template {
         it.properties = addOccurrence(intProperties, Occurrence)
-        it.inputFile = valueOverTimeAggregatorInputFile
-        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}OverTimeIntAggregator.java"
+        it.inputFile = valueByLongAggregatorInputFile
+        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}IntByLongAggregator.java"
       }
       template {
         it.properties = addOccurrence(longProperties, Occurrence)
-        it.inputFile = valueOverTimeAggregatorInputFile
-        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}OverTimeLongAggregator.java"
+        it.inputFile = valueByLongAggregatorInputFile
+        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}LongByLongAggregator.java"
       }
       template {
         it.properties = addOccurrence(floatProperties, Occurrence)
-        it.inputFile = valueOverTimeAggregatorInputFile
-        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}OverTimeFloatAggregator.java"
+        it.inputFile = valueByLongAggregatorInputFile
+        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}FloatByLongAggregator.java"
       }
       template {
         it.properties = addOccurrence(doubleProperties, Occurrence)
-        it.inputFile = valueOverTimeAggregatorInputFile
-        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}OverTimeDoubleAggregator.java"
+        it.inputFile = valueByLongAggregatorInputFile
+        it.outputFile = "org/elasticsearch/compute/aggregation/${Occurrence}DoubleByLongAggregator.java"
       }
     }
   }
