+++ x-pack/plugin/esql/compute/src/main/generated-src/org/elasticsearch/compute/aggregation/LastDoubleByLongAggregator.java
@@ -9,14 +9,14 @@ package org.elasticsearch.compute.aggregation;
 
 // begin generated imports
 import org.elasticsearch.common.util.BigArrays;
-import org.elasticsearch.common.util.LongArray;
+import org.elasticsearch.common.util.DoubleArray;
 import org.elasticsearch.common.util.LongArray;
 import org.elasticsearch.compute.ann.Aggregator;
 import org.elasticsearch.compute.ann.GroupingAggregator;
 import org.elasticsearch.compute.ann.IntermediateState;
 import org.elasticsearch.compute.data.Block;
 import org.elasticsearch.compute.data.BooleanBlock;
-import org.elasticsearch.compute.data.LongBlock;
+import org.elasticsearch.compute.data.DoubleBlock;
 import org.elasticsearch.compute.data.IntVector;
 import org.elasticsearch.compute.data.LongBlock;
 import org.elasticsearch.compute.operator.DriverContext;
@@ -25,19 +25,17 @@ import org.elasticsearch.core.Releasables;
 
 /**
  * A time-series aggregation function that collects the Last occurrence value of a time series in a specified interval.
- * This class is generated. Edit `X-ValueOverTimeAggregator.java.st` instead.
+ * This class is generated. Edit `X-ValueByLongAggregator.java.st` instead.
  */
 @GroupingAggregator(
-    { @IntermediateState(name = "timestamps", type = "LONG_BLOCK"), @IntermediateState(name = "values", type = "LONG_BLOCK") }
+    { @IntermediateState(name = "timestamps", type = "LONG_BLOCK"), @IntermediateState(name = "values", type = "DOUBLE_BLOCK") }
 )
-public class LastOverTimeLongAggregator {
+public class LastDoubleByLongAggregator {
     public static GroupingState initGrouping(DriverContext driverContext) {
         return new GroupingState(driverContext.bigArrays());
     }
 
-    // TODO: Since data in data_streams is sorted by `_tsid` and timestamp in descending order,
-    // we can read the first encountered value for each group of `_tsid` and time bucket.
-    public static void combine(GroupingState current, int groupId, long value, long timestamp) {
+    public static void combine(GroupingState current, int groupId, double value, long timestamp) {
         current.collectValue(groupId, timestamp, value);
     }
 
@@ -45,7 +43,7 @@ public class LastOverTimeLongAggregator {
         GroupingState current,
         int groupId,
         LongBlock timestamps, // stylecheck
-        LongBlock values,
+        DoubleBlock values,
         int otherPosition
     ) {
         // TODO seen should probably be part of the intermediate representation
@@ -54,7 +52,7 @@ public class LastOverTimeLongAggregator {
             long timestamp = timestamps.getLong(timestamps.getFirstValueIndex(otherPosition));
             int firstIndex = values.getFirstValueIndex(otherPosition);
             for (int i = 0; i < valueCount; i++) {
-                current.collectValue(groupId, timestamp, values.getLong(firstIndex + i));
+                current.collectValue(groupId, timestamp, values.getDouble(firstIndex + i));
             }
         }
     }
@@ -66,7 +64,7 @@ public class LastOverTimeLongAggregator {
     public static final class GroupingState extends AbstractArrayState {
         private final BigArrays bigArrays;
         private LongArray timestamps;
-        private LongArray values;
+        private DoubleArray values;
         private int maxGroupId = -1;
 
         GroupingState(BigArrays bigArrays) {
@@ -77,7 +75,7 @@ public class LastOverTimeLongAggregator {
             try {
                 timestamps = bigArrays.newLongArray(1, false);
                 this.timestamps = timestamps;
-                this.values = bigArrays.newLongArray(1, false);
+                this.values = bigArrays.newDoubleArray(1, false);
                 success = true;
             } finally {
                 if (success == false) {
@@ -86,7 +84,7 @@ public class LastOverTimeLongAggregator {
             }
         }
 
-        void collectValue(int groupId, long timestamp, long value) {
+        void collectValue(int groupId, long timestamp, double value) {
             boolean updated = false;
             if (groupId < timestamps.size()) {
                 // TODO: handle multiple values?
@@ -116,13 +114,13 @@ public class LastOverTimeLongAggregator {
         public void toIntermediate(Block[] blocks, int offset, IntVector selected, DriverContext driverContext) {
             try (
                 var timestampsBuilder = driverContext.blockFactory().newLongBlockBuilder(selected.getPositionCount());
-                var valuesBuilder = driverContext.blockFactory().newLongBlockBuilder(selected.getPositionCount())
+                var valuesBuilder = driverContext.blockFactory().newDoubleBlockBuilder(selected.getPositionCount())
             ) {
                 for (int p = 0; p < selected.getPositionCount(); p++) {
                     int group = selected.getInt(p);
                     if (group < timestamps.size() && hasValue(group)) {
                         timestampsBuilder.appendLong(timestamps.get(group));
-                        valuesBuilder.appendLong(values.get(group));
+                        valuesBuilder.appendDouble(values.get(group));
                     } else {
                         timestampsBuilder.appendNull();
                         valuesBuilder.appendNull();
@@ -134,11 +132,11 @@ public class LastOverTimeLongAggregator {
         }
 
         Block evaluateFinal(IntVector selected, GroupingAggregatorEvaluationContext evalContext) {
-            try (var builder = evalContext.blockFactory().newLongBlockBuilder(selected.getPositionCount())) {
+            try (var builder = evalContext.blockFactory().newDoubleBlockBuilder(selected.getPositionCount())) {
                 for (int p = 0; p < selected.getPositionCount(); p++) {
                     int group = selected.getInt(p);
                     if (group < timestamps.size() && hasValue(group)) {
-                        builder.appendLong(values.get(group));
+                        builder.appendDouble(values.get(group));
                     } else {
                         builder.appendNull();
                     }
