+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/aggregate/LastOverTime.java
@@ -10,10 +10,10 @@ package org.elasticsearch.xpack.esql.expression.function.aggregate;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.compute.aggregation.AggregatorFunctionSupplier;
-import org.elasticsearch.compute.aggregation.LastOverTimeDoubleAggregatorFunctionSupplier;
-import org.elasticsearch.compute.aggregation.LastOverTimeFloatAggregatorFunctionSupplier;
-import org.elasticsearch.compute.aggregation.LastOverTimeIntAggregatorFunctionSupplier;
-import org.elasticsearch.compute.aggregation.LastOverTimeLongAggregatorFunctionSupplier;
+import org.elasticsearch.compute.aggregation.LastDoubleByLongAggregatorFunctionSupplier;
+import org.elasticsearch.compute.aggregation.LastFloatByLongAggregatorFunctionSupplier;
+import org.elasticsearch.compute.aggregation.LastIntByLongAggregatorFunctionSupplier;
+import org.elasticsearch.compute.aggregation.LastLongByLongAggregatorFunctionSupplier;
 import org.elasticsearch.xpack.esql.EsqlIllegalArgumentException;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
 import org.elasticsearch.xpack.esql.core.expression.Literal;
@@ -128,12 +128,14 @@ public class LastOverTime extends TimeSeriesAggregateFunction implements Optiona
 
     @Override
     public AggregatorFunctionSupplier supplier() {
+        // TODO: When processing TSDB data_streams they are sorted by `_tsid` and timestamp in descending order,
+        // we can read the first encountered value for each group of `_tsid` and time bucket.
         final DataType type = field().dataType();
         return switch (type) {
-            case LONG -> new LastOverTimeLongAggregatorFunctionSupplier();
-            case INTEGER -> new LastOverTimeIntAggregatorFunctionSupplier();
-            case DOUBLE -> new LastOverTimeDoubleAggregatorFunctionSupplier();
-            case FLOAT -> new LastOverTimeFloatAggregatorFunctionSupplier();
+            case LONG -> new LastLongByLongAggregatorFunctionSupplier();
+            case INTEGER -> new LastIntByLongAggregatorFunctionSupplier();
+            case DOUBLE -> new LastDoubleByLongAggregatorFunctionSupplier();
+            case FLOAT -> new LastFloatByLongAggregatorFunctionSupplier();
             default -> throw EsqlIllegalArgumentException.illegalDataType(type);
         };
     }
